+++ src/hooks/useMediatedRequestById/useMediatedRequestById.js
@@ -1,3 +1,7 @@
+import {
+  useState,
+} from 'react';
+
 import { useQuery } from 'react-query';
 
 import {
@@ -6,13 +10,15 @@ import {
 } from '@folio/stripes/core';
 
 const useMediatedRequestById = (mediatedRequestId) => {
+  const [shouldUpdateMediatedRequestById, setShouldUpdateMediatedRequestById] = useState(0);
+
   const ky = useOkapiKy();
   const [namespace] = useNamespace({ key: 'mediatedRequest' });
   const {
     data,
     isFetching,
   } = useQuery(
-    [namespace, mediatedRequestId],
+    [namespace, mediatedRequestId, shouldUpdateMediatedRequestById],
     () => ky.get(`requests-mediated/mediated-requests/${mediatedRequestId}`).json(),
     { enabled: Boolean(mediatedRequestId) },
   );
@@ -20,6 +26,8 @@ const useMediatedRequestById = (mediatedRequestId) => {
   return {
     isFetching,
     mediatedRequest: data,
+    shouldUpdateMediatedRequestById,
+    setShouldUpdateMediatedRequestById,
   };
 };