+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Implement functionality for editing mediated request. Refs UIREQMED-22.
 * *BREAKING* Migrate to new `mod-circulation-bff` endpoints. Refs UIREQMED-39.
 * Add circulation-storage.staff-slips.collection.get for get staff-slips on Send item in transit. Refs UIREQMED-55.
+* Add Decline Action for Mediated request with status of New - Awaiting confirmation. Refs UIREQMED-42.
 
 ## [1.1.0](https://github.com/folio-org/ui-requests-mediated/tree/v1.1.0) (2024-10-30)
 * Update github actions. Refs UIREQMED-14.